% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map FARS Data for a Specific State and Year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer representing the state number.}

\item{year}{An integer/character specifying the year.}
}
\value{
An invisible NULL. The function generates a map of accidents for the specified state and year.
}
\description{
This function maps FARS (Fatality Analysis Reporting System) data for a specific state and year.
}
\details{
This function uses the maps package to plot accidents on a map for the specified state and year. If the state number
is invalid or if there are no accidents for the specified state and year, appropriate messages are displayed.
}
\examples{
# Map FARS data for state number 5 in the year 2013
# fars_map_state(5, 2013)

}
\seealso{
\code{\link{make_filename}}, \code{\link{fars_read}}
}
