% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS Data from CSV File}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The path to the CSV file.}
}
\value{
A tibble containing the data read from the CSV file.
}
\description{
This function reads data from a CSV file using the readr package and returns it as a tibble.
}
\details{
This function checks if the specified file exists before attempting to read it. If the file does not exist,
the function stops and displays an error message.
}
\examples{
# Read FARS data from a CSV file
# fars_data <- fars_read("path/to/your/file.csv")

}
