% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS Data for Multiple Years
This function reads FARS data for multiple years and returns a list of tibbles.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of integers/characters representing the years to read data for.}
}
\value{
A list of tibbles, each containing FARS data for a specific year.
}
\description{
Read FARS Data for Multiple Years
This function reads FARS data for multiple years and returns a list of tibbles.
}
\details{
This function uses the fars_read function to read data for each specified year with the corresponding months. If an invalid year is encountered,
a warning is issued, and the corresponding data for that year is set to NULL in the output list.
}
\examples{
# Read FARS data for the years 2013 and 2014
# data_list <- fars_read_years(c(2013, 2014))

}
\seealso{
\code{\link{make_filename}}, \code{\link{fars_read}}
}
