% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS Data for Multiple Years
This function summarizes FARS (Fatality Analysis Reporting System) data for multiple years by counting the number of accidents
for each month in each year.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of integers/characters representing the years to summarize data for.}
}
\value{
A tibble containing the summarized data with columns for the year, month, and the number of accidents.
}
\description{
Summarize FARS Data for Multiple Years
This function summarizes FARS (Fatality Analysis Reporting System) data for multiple years by counting the number of accidents
for each month in each year.
}
\examples{
# Summarize FARS data for the years 2013 and 2014
# summary_data <- fars_summarize_years(c(2013, 2014))

}
\seealso{
\code{\link{fars_read_years}}
}
