% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Generate FARS Data Filename
This function generates a filename for FARS (Fatality Analysis Reporting System) data for a given year.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer/character specifying the year.}
}
\value{
A character string representing the filename.
}
\description{
Generate FARS Data Filename
This function generates a filename for FARS (Fatality Analysis Reporting System) data for a given year.
}
\examples{
# Generate a filename for FARS data for the year 2013
# filename <- make_filename(2021)
# Output: "accident_2021.csv.bz2"

}
